# SALTO-6307: Improve GroupMembership fetch in Okta