--- a/packages/okta-adapter/src/adapter.ts
+++ b/packages/okta-adapter/src/adapter.ts
@@ -197,12 +197,15 @@ export default class OktaAdapter implements AdapterOperations {
       client: this.client,
       paginationFuncCreator: paginate,
     })
+    this.fetchQuery = elementUtils.query.createElementQuery(this.userConfig.fetch, fetchCriteria)
+
     const definitions = {
       // TODO - SALTO-5746 - only provide adminClient when it is defined
       clients: createClientDefinitions({ main: this.client, private: this.adminClient ?? this.client }),
       pagination: PAGINATION,
       fetch: createFetchDefinitions(
         this.userConfig,
+        this.fetchQuery,
         shouldAccessPrivateAPIs(this.isOAuthLogin, this.userConfig),
         getAdminUrl(this.client.baseUrl),
       ),
@@ -217,8 +220,6 @@ export default class OktaAdapter implements AdapterOperations {
         fetchConfig: definitions.fetch,
       }),
     }
-
-    this.fetchQuery = elementUtils.query.createElementQuery(this.userConfig.fetch, fetchCriteria)
 
     this.paginator = paginator