--- a/packages/okta-adapter/src/adapter.ts
+++ b/packages/okta-adapter/src/adapter.ts
@@ -203,12 +203,12 @@ export default class OktaAdapter implements AdapterOperations {
       // TODO - SALTO-5746 - only provide adminClient when it is defined
       clients: createClientDefinitions({ main: this.client, private: this.adminClient ?? this.client }),
       pagination: PAGINATION,
-      fetch: createFetchDefinitions(
-        this.userConfig,
-        this.fetchQuery,
-        shouldAccessPrivateAPIs(this.isOAuthLogin, this.userConfig),
-        getAdminUrl(this.client.baseUrl),
-      ),
+      fetch: createFetchDefinitions({
+        userConfig: this.userConfig,
+        fetchQuery: this.fetchQuery,
+        usePrivateAPI: shouldAccessPrivateAPIs(this.isOAuthLogin, this.userConfig),
+        baseUrl: getAdminUrl(this.client.baseUrl),
+      }),
       deploy: createDeployDefinitions(),
       sources: { openAPI: [OPEN_API_DEFINITIONS] },
     }