--- a/packages/okta-adapter/src/filters/group_members.ts
+++ b/packages/okta-adapter/src/filters/group_members.ts
@@ -15,15 +15,7 @@
  */
 import _ from 'lodash'
 import {
-  Element,
   InstanceElement,
-  isInstanceElement,
-  CORE_ANNOTATIONS,
-  ReferenceExpression,
-  ObjectType,
-  ElemID,
-  BuiltinTypes,
-  ListType,
   isAdditionOrModificationChange,
   isInstanceChange,
   getChangeData,
@@ -32,19 +24,15 @@ import {
   isAdditionChange,
   SaltoElementError,
 } from '@salto-io/adapter-api'
-import { elements as elementUtils, client as clientUtils } from '@salto-io/adapter-components'
-import { pathNaclCase, safeJsonStringify, applyFunctionToChangeData, getParents } from '@salto-io/adapter-utils'
-import { collections, values } from '@salto-io/lowerdash'
+import { client as clientUtils } from '@salto-io/adapter-components'
+import { safeJsonStringify, applyFunctionToChangeData, getParents } from '@salto-io/adapter-utils'
+import { values } from '@salto-io/lowerdash'
 import { logger } from '@salto-io/logging'
 import { FilterCreator } from '../filter'
-import { GROUP_TYPE_NAME, GROUP_MEMBERSHIP_TYPE_NAME, OKTA } from '../constants'
-import { areUsers, shouldConvertUserIds, User } from '../user_utils'
+import { GROUP_MEMBERSHIP_TYPE_NAME } from '../constants'
 import { FETCH_CONFIG } from '../config'
 
 const log = logger(module)
-const { RECORDS_PATH, TYPES_PATH } = elementUtils
-const { toArrayAsync } = collections.asynciterable
-const { makeArray } = collections.array
 const { isDefined } = values
 
 type GroupMembershipInstance = InstanceElement & {
@@ -56,54 +44,6 @@ type GroupMembershipInstance = InstanceElement & {
 type GroupMembershipDeployResult = {
   appliedChange?: AdditionChange<InstanceElement> | ModificationChange<InstanceElement>
   error?: SaltoElementError
-}
-
-const createGroupMembershipType = (): ObjectType =>
-  new ObjectType({
-    elemID: new ElemID(OKTA, GROUP_MEMBERSHIP_TYPE_NAME),
-    fields: {
-      members: { refType: new ListType(BuiltinTypes.STRING) },
-    },
-    path: [OKTA, TYPES_PATH, GROUP_MEMBERSHIP_TYPE_NAME],
-  })
-
-const getGroupMembersData = async (paginator: clientUtils.Paginator, group: InstanceElement): Promise<User[]> => {
-  const paginationArgs = {
-    url: `/api/v1/groups/${group.value.id}/users`,
-    paginationField: 'after',
-  }
-  const members = (
-    await toArrayAsync(paginator(paginationArgs, page => makeArray(page) as clientUtils.ResponseValue[]))
-  ).flat()
-  if (!areUsers(members)) {
-    log.error(`Received invalid response while trying to get members for group: ${group.elemID.getFullName()}`)
-    return []
-  }
-  return members
-}
-
-const createGroupMembershipInstance = async ({
-  group,
-  groupMembersType,
-  paginator,
-  userIdentifier,
-}: {
-  group: InstanceElement
-  groupMembersType: ObjectType
-  paginator: clientUtils.Paginator
-  userIdentifier: 'id' | 'email'
-}): Promise<InstanceElement | undefined> => {
-  const groupName = group.elemID.name
-  const groupMembersData = await getGroupMembersData(paginator, group)
-  return groupMembersData.length > 0
-    ? new InstanceElement(
-        groupName,
-        groupMembersType,
-        { members: groupMembersData.map(member => (userIdentifier === 'id' ? member.id : member.profile.login)) },
-        [OKTA, RECORDS_PATH, GROUP_MEMBERSHIP_TYPE_NAME, pathNaclCase(groupName)],
-        { [CORE_ANNOTATIONS.PARENT]: [new ReferenceExpression(group.elemID, group)] },
-      )
-    : undefined
 }
 
 export const isValidGroupMembershipInstance = (instance: InstanceElement): instance is GroupMembershipInstance =>
@@ -216,33 +156,10 @@ const deployGroupMembershipChange = async (
 }
 
 /**
- * Create a single group-memberships instance per group.
+ * Deploy GroupMembership by adding or removing users from groups
  */
-const groupMembersFilter: FilterCreator = ({ definitions, config, paginator, fetchQuery }) => ({
+const groupMembersFilter: FilterCreator = ({ definitions, config }) => ({
   name: 'groupMembersFilter',
-  onFetch: async (elements: Element[]): Promise<void> => {
-    if (!config[FETCH_CONFIG].includeGroupMemberships) {
-      log.debug('Fetch of group members is disabled')
-      return
-    }
-    const groupInstances = elements
-      .filter(isInstanceElement)
-      .filter(instance => instance.elemID.typeName === GROUP_TYPE_NAME)
-
-    const groupMembersType = createGroupMembershipType()
-    elements.push(groupMembersType)
-
-    const userIdentifier = shouldConvertUserIds(fetchQuery, config) ? 'email' : 'id'
-    const groupMembershipInstances = (
-      await Promise.all(
-        groupInstances.map(async group =>
-          createGroupMembershipInstance({ group, groupMembersType, paginator, userIdentifier }),
-        ),
-      )
-    ).filter(isInstanceElement)
-
-    groupMembershipInstances.forEach(instance => elements.push(instance))
-  },
   deploy: async changes => {
     const client = definitions.clients.options.main.httpClient
     const [relevantChanges, leftoverChanges] = _.partition(