--- a/packages/okta-adapter/tsconfig.json
+++ b/packages/okta-adapter/tsconfig.json
@@ -4,7 +4,7 @@
     "outDir": "dist",
     "baseUrl": "."
   },
-  "include": ["src/**/*", "src/**/*.json", "index.ts", "e2e_test/jest_environment.ts"],
+  "include": ["src/**/*", "src/**/*.json", "index.ts", "e2e_test/jest_environment.ts", "test/utils.ts"],
   "references": [
     { "path": "../adapter-api" },
     { "path": "../adapter-components" },