… packages/okta-adapter/test/fetch_mock_replies.json