Attempt to optimize fetch times for `GroupMembership` instance

---

_Additional context for reviewer_

The filter responsible for fetching group members took pretty long time in envs with many groups (as we make a request per group), I made some changes that hopefully would improve performance:
1. Moved `GroupMembership` type to be fetched by infra instead of a filter, so there won’t be a tail of many requests fetching the group members when the filters run
2. When fetching groups, I added a query param that adds the response the number of users assigned to the group. We then temporarily store the result on the instance to use with conditional `recurseInto`. This way we only make a request for groups that has any members assigned.

I did some measurements in our big okta env - 500 groups, 2000 users, group memberships only in some groups
(this doesn't represent good enough actual env cause not all groups has assignments :|)


| Before (ms) | After (ms) |
| -------- | --- |
| 1010644 |  259287 |
|1167562 | 259248 |
| - | 258331 |



---
_Release Notes_: 

_Okta_adapter_:
- Improve fetch time when including group memberships

---
_User Notifications_: 
None